Move fragment configurations to shared object
